move all comments block effects into provider
