+++ js/codebook.toml
@@ -1,0 +1,1 @@
+words = ["loro"]
